Rename helpers -> attestation_parent_hashes
